[REMOVE] 키워드 관련 로직 삭제
기능 요구사항의 변경으로 전체 코드에서 키워드 관련 로직 삭제
