+++ src/main/java/com/wss/websoso/novel/NovelController.java
@@ -1,17 +1,13 @@
 package com.wss.websoso.novel;
 
-import com.wss.websoso.userNovel.UserNovelCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
 import java.security.Principal;
 import java.util.List;
 
@@ -37,19 +33,5 @@ public class NovelController {
     @GetMapping("{novelId}")
     public ResponseEntity getNovelByNovelId(@PathVariable Long novelId, Principal principal) {
         return novelService.getNovelByNovelId(novelId, Long.valueOf(principal.getName()));
-    }
-
-    @PostMapping("{novelId}")
-    public ResponseEntity<Void> createUserNovel(
-            @PathVariable Long novelId,
-            @RequestBody UserNovelCreateRequest userNovelCreateRequest,
-            Principal principal) {
-        URI location = URI.create("/userNovels/" + novelService.createUserNovel(
-                novelId,
-                Long.valueOf(principal.getName()),
-                userNovelCreateRequest)
-        );
-
-        return ResponseEntity.created(location).build();
     }
 }
