+++ src/main/java/com/wss/websoso/novel/NovelService.java
@@ -1,19 +1,12 @@
 package com.wss.websoso.novel;
 
-import com.wss.websoso.config.ReadStatus;
-import com.wss.websoso.keyword.Keyword;
-import com.wss.websoso.keyword.KeywordRepository;
-import com.wss.websoso.platform.Platform;
 import com.wss.websoso.platform.PlatformGetResponse;
 import com.wss.websoso.platform.PlatformRepository;
 import com.wss.websoso.user.User;
 import com.wss.websoso.user.UserRepository;
 import com.wss.websoso.userNovel.UserNovel;
-import com.wss.websoso.userNovel.UserNovelCreateRequest;
 import com.wss.websoso.userNovel.UserNovelGetResponse;
 import com.wss.websoso.userNovel.UserNovelRepository;
-import com.wss.websoso.userNovelKeyword.UserNovelKeyword;
-import com.wss.websoso.userNovelKeyword.UserNovelKeywordRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -21,7 +14,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,8 +23,6 @@ public class NovelService {
     private final NovelRepository novelRepository;
     private final UserRepository userRepository;
     private final UserNovelRepository userNovelRepository;
-    private final UserNovelKeywordRepository userNovelKeywordRepository;
-    private final KeywordRepository keywordRepository;
     private final PlatformRepository platformRepository;
 
     public List<NovelGetResponse> getNovelsByWord(Long lastNovelId, int size, String word) {
@@ -71,9 +61,6 @@ public class NovelService {
                     userNovel.getUserNovelReadStatus(),
                     userNovel.getUserNovelReadStartDate(),
                     userNovel.getUserNovelReadEndDate(),
-                    userNovel.getUserNovelKeywords().stream()
-                            .map(userNovelKeyword -> userNovelKeyword.getKeyword().getKeywordName())
-                            .toList(),
                     userNovel.getPlatforms().stream()
                             .map(platform -> new PlatformGetResponse(
                                     platform.getPlatformName(),
@@ -97,63 +84,5 @@ public class NovelService {
                             .toList()
             ));
         }
-    }
-
-    public Long createUserNovel(Long novelId, Long userId, UserNovelCreateRequest userNovelCreateRequest) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("해당하는 사용자가 없습니다."));
-
-        Novel novel = novelRepository.findById(novelId)
-                .orElseThrow(() -> new IllegalArgumentException("해당하는 작품이 없습니다."));
-
-        UserNovel userNovel = UserNovel.builder()
-                .novelId(novelId)
-                .userNovelTitle(novel.getNovelTitle())
-                .userNovelAuthor(novel.getNovelAuthor())
-                .userNovelGenre(novel.getNovelGenre())
-                .userNovelImg(novel.getNovelImg())
-                .userNovelDescription(novel.getNovelDescription())
-                .userNovelRating(userNovelCreateRequest.novelRating())
-                .userNovelReadStatus(ReadStatus.valueOf(userNovelCreateRequest.novelReadStatus()))
-                .userNovelReadStartDate(userNovelCreateRequest.novelReadStartDate())
-                .userNovelReadEndDate(userNovelCreateRequest.novelReadEndDate())
-                .user(user)
-                .build();
-
-        userNovelRepository.save(userNovel);
-
-        if (novel.getPlatforms() != null) {
-            novel.getPlatforms().stream()
-                    .forEach(platform -> {
-                        Platform newPlatform = Platform.builderWithUserNovel()
-                                .platformName(platform.getPlatformName())
-                                .platformUrl(platform.getPlatformUrl())
-                                .userNovel(userNovel)
-                                .buildWithUserNovel();
-
-                        platformRepository.save(newPlatform);
-                    });
-        }
-
-        if (userNovelCreateRequest.keywordNames() != null) {
-            userNovelCreateRequest.keywordNames().stream()
-                    .map(keywordName -> {
-                        Optional<Keyword> optionalKeyword = keywordRepository.findByKeywordName(keywordName);
-                        if (optionalKeyword.isEmpty()) {
-                            throw new IllegalArgumentException("해당하는 키워드가 없습니다.");
-                        }
-                        return optionalKeyword.get();
-                    })
-                    .forEach(keyword -> {
-                        UserNovelKeyword userNovelKeyword = UserNovelKeyword.builder()
-                                .userNovel(userNovel)
-                                .keyword(keyword)
-                                .build();
-
-                        userNovelKeywordRepository.save(userNovelKeyword);
-                    });
-        }
-
-        return userNovel.getUserNovelId();
     }
 }
