+++ src/main/java/com/wss/websoso/userNovel/UserNovelController.java
@@ -1,16 +1,20 @@
 package com.wss.websoso.userNovel;
 
 import com.wss.websoso.config.ReadStatus;
-import java.security.Principal;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.security.Principal;
+import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
@@ -18,6 +22,20 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserNovelController {
 
     private final UserNovelService userNovelService;
+
+    @PostMapping("{novelId}")
+    public ResponseEntity<Void> createUserNovel(
+            @PathVariable Long novelId,
+            @RequestBody UserNovelCreateRequest userNovelCreateRequest,
+            Principal principal) {
+        URI location = URI.create("/userNovels/" + userNovelService.createUserNovel(
+                novelId,
+                Long.valueOf(principal.getName()),
+                userNovelCreateRequest)
+        );
+
+        return ResponseEntity.created(location).build();
+    }
 
     @GetMapping
     public UserNovelsResponse getUserNovels(
